:recycle: Refactor: 도메인 로직 이동
<body>
SettlementCalculator는 도메인 로직에 해당하기 때문에 db.domain 모듈로 이동

- 관련 : #343
